Editor: Update notice action to in app preview
When a new page or post is published/scheduled/updated, we display a success
notice including a link to open the post/page. This updates the notice action in
that notice to use the in app preview if available instead of opening in another
tab. The commit also includes tests for the desired NoticeAction behavior.