My Sites: Improve speed of switch from Reader to My Sites
When user switches between the Reader and My Sites there is a fairly long pause, this commit tries to get that pause bellow 250ms so it will perceive instant.

It achieves it by extending a component of My Sites to be pure and thus eliminates wasted render calls when the data those components rely on do no change.