wip - implement sites observer hoc
using Object.create to create a new instance of object with the properties and methods of the original SitesList by using it as prototype.

Also add tiny check in sites-list/list to eliminate emit of change event when selecting the same site. This happens when navigating inside My Sites between stats, pages and posts for example.